Update data structures to support CTF global type declarations
The structures match those in LTTng-tools to support global type
declarations in general and CTF enumerations in particular.

Note: This patch uses the lttng_enum structure in ust-events.h and
modifies the container_type field. This structure was not used and has
never been used as far as we know.

[ Edit by Mathieu Desnoyers: fix the padding, especially for existing
  union (should not change), and add missing LTTNG_PACKED. ]

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>