Send and receive serialized data for CTF global enumerations
Updates the functions used for the communication between UST and the
session daemon so that global type declarations can be sent and
received. The description of the global type declarations follows the
event that uses them.  The extra fields for each global type category
follow the descriptions of all global type declarations.

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>