Introduce LTTng-UST perf sampling
Perf can be configured to trigger a signal when the performance counter
reaches a given value. Inside the signal handler, we save the call stack
using libunwind, which works without frame pointers. The sample is saved to
LTTng-UST tracepoint as an ctf_sequence.

The instrumentation is contained in a shared library, such that the
profiling is enabled when the library is loaded. It would also be possible
to export functions to control the sampling from within the profiled
program.

The overhead for a cpu-bound computation with the default sampling rate is
about:

 baseline: 2.56s
 tracing : 2.80s
 overhead: 0.24s (8.5%)

The event cost on average is 2.2us.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>